datamodel_generator/requirements.txt
PyYAML
Jinja2
streamlit
pyyaml
psycopg2
graphviz
